+++ choco/src/main/java/btrplace/solver/choco/transition/KillVM.java
@@ -36,7 +36,7 @@ import java.util.EnumSet;
 /**
  * An action to model a VM that is killed.
  * An estimation of the action duration must be provided through a
- * {@link btrplace.solver.choco.durationEvaluator.ActionDurationEvaluator} accessible from
+ * {@link btrplace.solver.choco.duration.ActionDurationEvaluator} accessible from
  * {@link btrplace.solver.choco.ReconfigurationProblem#getDurationEvaluators()} with the key {@code KillVM.class}
  * <p/>
  * If the reconfiguration problem has a solution, a {@link btrplace.plan.event.KillVM} action