+++ choco/src/main/java/btrplace/solver/choco/transition/ShutdownVM.java
@@ -32,7 +32,7 @@ import solver.variables.VariableFactory;
 /**
  * Model an action that stop a running VM.
  * An estimation of the action duration must be provided through a
- * {@link btrplace.solver.choco.durationEvaluator.ActionDurationEvaluator} accessible from
+ * {@link btrplace.solver.choco.duration.ActionDurationEvaluator} accessible from
  * {@link btrplace.solver.choco.ReconfigurationProblem#getDurationEvaluators()} with the key {@code ShutdownVM.class}
  * <p/>
  * If the reconfiguration problem has a solution, a {@link btrplace.plan.event.ShutdownVM} action is inserted