+++ examples/src/main/java/btrplace/examples/ModelCustomization.java
@@ -27,9 +27,9 @@ import btrplace.plan.event.ShutdownVM;
 import btrplace.solver.SolverException;
 import btrplace.solver.choco.ChocoReconfigurationAlgorithm;
 import btrplace.solver.choco.DefaultChocoReconfigurationAlgorithm;
-import btrplace.solver.choco.durationEvaluator.ConstantActionDuration;
-import btrplace.solver.choco.durationEvaluator.DurationEvaluators;
-import btrplace.solver.choco.durationEvaluator.LinearToAResourceActionDuration;
+import btrplace.solver.choco.duration.ConstantActionDuration;
+import btrplace.solver.choco.duration.DurationEvaluators;
+import btrplace.solver.choco.duration.LinearToAResourceActionDuration;
 
 import java.util.ArrayList;
 import java.util.Arrays;