+++ examples/src/main/java/btrplace/examples/SolverTuning.java
@@ -27,7 +27,7 @@ import btrplace.plan.event.MigrateVM;
 import btrplace.solver.SolverException;
 import btrplace.solver.choco.ChocoReconfigurationAlgorithm;
 import btrplace.solver.choco.DefaultChocoReconfigurationAlgorithm;
-import btrplace.solver.choco.durationEvaluator.LinearToAResourceActionDuration;
+import btrplace.solver.choco.duration.LinearToAResourceActionDuration;
 
 import java.util.*;